requirements.txt
pandas
numpy>=1.23 
gcloud
dvc
setuptools
pyyaml
tensorflow
google-cloud-storage
dvc-gs
flask
joblib
scikit-learn
comet-ml